Report detailed errors if port requests cannot be handled
Instead of failing silently when PortProvider receives a message that
has the correct structure for a port request but fails validation of the
message details or origin, report an error. This will make debugging
these issues easier both locally and in production, where the error will
be captured and reported to Sentry.