Ignore unknown keys when checking payload / message match
When comparing a MessageEvent's data to an expected message, only
compare the known fields of the message and ignore any unknown fields.

Unknown fields don't cause problems for us in parsing the message, and
there is a risk that it might cause us to silently ignore messages in
some environments. We don't know yet if that actually happens.