Fix race condition if annotations are deleted before anchoring
This commit fixes a situation when annotations are deleted in the
`sidebar` frame before the `guest` frame finishes to anchor the
annotations.

`Guest` tracks of which annotations needs to be deleted in a field
`Guest#_deferredDetach`. This annotation set is check on every anchor
event.

`Guest` also adds a `Map` that tracks of which annotations have been
anchored. This `Guesst#_annotations` field allows for future performance
optimisations.