Rebuild boot script when modules in src/boot change
Previously the boot script was not regenerated when its JS inputs
changed, as opposed to the JSON manifest.

This approach relies on `src/boot` not having dependencies on other
modules. We could improve this by using `watchJS`, but that currently
produces too much distracting log noise when running `make dev`.