Handle port requests from guests with opaque origins
When a guest frame with an opaque origin, such as file:// URL or a sandboxed
iframe, sends a request for a port to the host frame, the `event.origin` value
will be the string "null". This is not a legal value for the `targetOrigin`
argument to `postMessage`, so we need to remap it to "*" (any origin).