Remove a test which doesn't test what it intends to
Dates are in fact JSON-serializable, but as strings.

Recent changes to `isMessageEqual` mean that the `JSON.stringify` calls
cannot fail, because the `isMessage` check ensures all the fields we
serialize are strings. Writing a test to check for this failure mode is
not possible. There is a refactoring hazard here, so I added a comment
to call this out.