Clean up `Guest` tests
Created two functions `makeAnnotation` and `makeAnchor` that eliminates
a fair amount of boiler plate code in the test. `makeAnnotation` returns
an object of `AnnotationData` shape that can be feed to the
`Guest#anchor` method. `makeAnchor` simulates `Guest#anchor` and
allows to easily populates `Guest#anchors` field.

I moved `emitSidebarEvent` to the top.

Replaced all instances of `sinon` to `sandbox`.

Removed `Guest#focusedAnnotationTags` which was only used for testing
and it is not needed any more.

Simplified and standardized some tests.