Upload sourcemaps as part of Jenkins build
In fixing https://github.com/hypothesis/client/issues/3681, we will be
stripping parts of the script filename in stacktraces sent to Sentry. As
a result it will no longer be possible for Sentry to fetch the script
and associated sourcemap from https://cdn.hypothes.is/. Instead we need
to upload sourcemaps directly to Sentry as part of the build process.

See https://docs.sentry.io/platforms/javascript/sourcemaps/uploading/