Create an dev test scenario for a sandboxed guest frame
This will allow us to test the effects of sandboxing the guest iframes.
In the past, it affected the inter-frame communication (#4012) but can
also affect other process. For example, it has revealed a possible
collision with the `createIntegration` function:

```
Sandbox access violation: Blocked a frame at "http://localhost:3002" from accessing a frame at "http://localhost:3000".  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
```

The above error is from Safari.