Remove all comments from minified JavaScript
By default Terser keeps certain comments [1] in minified output. Safari has a
bug [2] that causes column numbers to be incorrectly reported in stack
traces when a line contains inline comments. In minified code where the
output is packed into a small number of lines, this can result in stack
traces that map to the wrong source line. Stripping all comments from
the output works around the issue.

The effect on the size of the bundles is minimal. sidebar.bundle.js is
about 2KB smaller with this change.

Part of https://github.com/hypothesis/client/issues/4045

[1] https://github.com/terser/terser#format-options
[2] https://bugs.webkit.org/show_bug.cgi?id=221548