Reduce boot script cache expiry time
Reduce the potential for mismatches between the client versions loaded
in the annotator and sidebar when rolling out a new release by
shortening the cache expiry time on the boot script.

Set a very short expiry time on the browser cache TTL and a slightly
longer time on the edge cache.