Fix call to `store.focusGroup` in group-filtering test
The test was erroneously passing a Group object instead of a Group.id
and was causing warnings to be logged to the test output.

The immediately-following test passed because the group it was
attempting to focus was the group that would be focused by default anyway.
I adjusted the groups being focused and tested here to not use the
default-focused group to assure the test is testing what it means to.