WIP; proof of concept for notebook-sidebar communication
This PR enables a new communication channel between the notebook and the
sidebar. This in turn prevents the notebook from having to re-initialise
the notebook app (re-rendering parent iframe) each time the notebook is
opened, hence speeding up the app after the first rendering.

This is very crude WIP PR, just to demonstrate the feasibility of the
channel and to get preliminary feedback.