Wrap embedded YouTube videos in extension
Some YouTube embedded videos fail to load if no `Referer` header is sent
with the request [1]. This is the case when the sidebar or notebook are
served from the Chrome browser extension, as Chrome does not set a
`Referer` header when loading subresources for HTML documents loaded
from chrome-extension:// URLs.

Chrome does not provide an easy way to add the `Referer` header that
we've found, at least not without requesting scary-sounding additional permissions.
Implement a workaround which wraps YouTube videos in an iframe served
from https://cdn.hypothes.is/media-embed.html. This results in the video
player being requested with a `Referer: https://cdn.hypothes.is` header.

Fixes https://github.com/hypothesis/product-backlog/issues/1297.

[1] Specifically, it seems that YouTube videos containing copyrighted
    content are likely to be affected.