Fix error when VitalSource integration injects client into content frame
Construction of the integration was recently moved to happen earlier in the
Guest.  This caused a regression in the VitalSource integration, which may
synchronously call `Guest#injectClient` when constructed. `injectClient` in turn
relies on `_hypothesisInjector` being initialized, which was not the case.

The error could be seen in the dev server's VS test case at
http://localhost:3000/document/vitalsource-epub.

Fix the issue by moving `_hypothesisInjector` initialization to happen earlier
in the Guest constructor, before `createIntegration` is called.