Highlight the annotation's anchor on mouseover
Previously, when a user hover on an annotation it didn't change the
color of the anchor.

With this change, hovering an annotation result in the anchors to be
highlighted. The UI behaviour is therefore identical when hovering on an
annotation card in the sidebar, on an bucket in the bucket bar, or on an
annotation in the actual text.