Remove obsolete reference to `__hypothesis` global
The `hasHypothesis` function always returned false because it relied on a global
variable that no longer exists. The fact that it was broken didn't cause
noticeable problems because the client's boot script also has a check for
whether Hypothesis is already injected.

Replace the broken check with a working one that corresponds to the nearby
`injectHypothesis` function.