Add `yarn typecheck-no-any` command
Add a command which runs TypeScript against a subset of the codebase, currently
just src/shared/, with the `noImplicitAny` option enabled.

This will allow us to gradually work towards the goal of having the entire
client codebase typecheck with this option enabled. Once this state is reached
the `noImplicitAny` option can be enabled in the main TS config file
(src/tsconfig.json) and the `yarn typecheck-no-any` command and associated TS
config file can be removed.