Use same `Guest._frameIdentifier` as assigned by `PortProvider`
`Guest` class relies on a frame identifier set by the injector to
uniquely identify itself when communicating with the sidebar frame. This
identifier can have two types: `null`, when the hypothesis client was
initially loaded, and `string`, when the client is injected.

In this PR, the `Guest` waits until `PortFinder` to return a `requestId`
set by `PortProvider`. This `requestId` can be used to uniquely identify
guest frame during its lifetime.

This change allows to remove a mechanism in `FrameSync` that replaces a
temporal ID by a permanent one that comes with the metadata information
in the `documentInfoChanged` RPC event.