Wrap operations that may trigger component rerender in `act`
This may fix a flakey test failure that appears to be caused by
re-renders in these tests happening after the test run finishes, by
which time some of the necessary mocks have been undone.

Another thing we could do would be to explicitly unmount components at
the end of the test.