Use ResizeObserver to detect PDF image size changes
This allows the client to detect, in most supported browsers, changes in the
size and position of the PDF book image which don't affect the window
size. This will be needed to support side-by-side mode for VS PDF books.
It may also help if VS resizes the book image for any other reason.

For Safari < 13.4, we fall back to using window resize events.