Support side-by-side mode in VitalSource PDF books
Support side-by-side mode in PDF books by simply resizing the content of
the page. This is similar to how side-by-side mode works in PDFs.

The width of the page in the VS PDF demo has been increased so that it
better matches the actual VS viewer, and also so that it triggers the
heuristics in the Guest class for detecting an iframe that fills the
parent frame.