Re-implement `getScrollAnchor` to ignore text nodes in hidden elements
The previous implementation of `getScrollAnchor` could choose words from
a non-displayed text node (eg. inside a `<title>` tag) as the anchor.
This is because `Range.getBoundingClientRect()` for such a word would
return a DOMRect with coordinates (0, 0, 0, 0) which passed the test for
being contained within the viewport (0, 0, window.innerWidth,
window.innerHeight).

Fix this with a rewrite of `getScrollAnchor` which fixes this problem by
skipping over elements that are hidden or have zero size. The new method
is also more efficient as it skips over text nodes in subtrees of the
root element which do not intersect the viewport.

Various utility method related to testing the emptiness and relation
between DOMRects have been split out into a separate module for ease of
testing and also because they are likely to come in useful for other
parts of the code.