Enable side-by-side mode for iframe guests that fill the host page
To make side-by-side mode work for VitalSource EPUB books, the guest in
the VS content iframe needs to invoke the integration's `fitSideBySide`
method. This is done in a generic way by making guests in iframes call
this method, but only if they know that the iframe fills the host frame.