Limit VitalSource iframe height in PDFs
VitalSource PDF books currently create very tall iframes. When a user
scrolls through a PDF page, they are scrolling the _parent_ frame rather
than the iframe itself. The bucket bar and scrolling logic in the client
don't support this. The result is that:

 - Bucket bar buckets are in the wrong location and don't scroll as the
   user scrolls the parent frame
 - Clicking on an annotation / bucket tries to scroll only the iframe,
   not the parent frame.

To resolve this we need to do at least one of:

1. Support these tall iframes in the client generally
2. Forcibly change the height of the iframe and enable the iframe to scroll
3. Persuade VitalSource to change the iframe layout so that the content frame
   scrolls, rather than the parent frame. They previously did this for EPUBs
   already.

This commit implements a version of (2). If VS later apply (3) then we can
remove this workaround.