Take into account padding when computing free space
When computing the amount of free space between the right edge of the
main content and the edge of the sidebar, allow for the padding
allocated between the sidebar and the content. This fixes an issue where
the main content would touch the sidebar exactly instead of having a
margin between it and the content.

Also tweak the padding value slightly to be a multiple of the 4px grid
cell size used in the design system.