Make side-by-side mode work with VitalSource "Great Book" books
Many classic texts that are freely available in VitalSource use `<div
class="para">` elements for paragraphs of text, instead of `<p>` elements.
Modify the logic for finding the main content of an HTML document to consider
these. We will probably want to make these selectors a configurable argument to
`guessMainContentArea` in future. Since there are only two of them and it is
unlikely to have ill effects outside of VitalSource books, I made it a general
feature of the HTML side-by-side mode implementation.