Create `scrollElementIntoView` wrapper around smooth-scrolling logic
Extract the logic for smoothly scrolling an element into view into the
scrolling utility module, in preparation for adding a workaround for an
issue with XHTML documents. This will also provide a place to configure
common defaults and enable us to swap out the underlying implementation
more easily in future.