Work around scroll-into-view issue in XHTML documents
Add a workaround for the scroll-into-view package for XHTML documents, including
EPUBs. In these documents `scrollIntoView` failed to detect the <body> element
correctly when finding scrollable ancestors of the target element. As a result
the client did not scroll to the correct location for a highlight when
clicking on annotation cards or bucket bar items.

The workaround is to make `document.body.tagName` behave the same in XHTML
documents as it does in HTML documents, and rely on this being unlikely to break
any of the page's own JavaScript.