Increase timeout in test
`waitFor` has a short default timeout of 10ms and this test sometimes
failed in CI. Try increasing it to reduce flakiness. We might want to
consider changing the `waitFor` default in future, although we also want
to encourage test authors to avoid writing tests that frequently wait
for long (> 10ms) periods of time.