Refactor `createStore` helpers
Move helper functions used only by `createStore` into `create-store.js` and
test the functionality as part of the `createStore` tests rather than having
direct tests for the helpers.

This refactoring decouples the tests for `createStore` from implementation
details. Having a single module for `createStore` and its tests should also
make it easier for a new developer to grok it.

 - Move `createReducer` and `bindSelectors` from util.js into create-store.js and
   refactor `bindSelectors`

 - Remove dedicated tests for these helpers and instead test the functionality
   via tests for `createStore`.

 - Remove check for `namespace` property of modules, as the types for the
   `createStoreModule` function now enforce the presence of this property.

 - Remove checks for conflicting selectors. These will be re-added in a
   subsequent commit.