Build and load annotator bundle as a non-module script
Safari does not support loading module scripts in XHTML pages that are served
with the "application/xhtml" mime type [1]. This is very rare on the open web (XHTML
is rare and XHTML documents served with "application/xhtml" as opposed to
"text/html" as the MIME type are rarer still [2]) but more common in ebooks.

This partly fixes https://github.com/hypothesis/client/issues/4350.

[1] https://bugs.webkit.org/show_bug.cgi?id=227469
[2] https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/XHTML#xhtml_document