Fix crash if link href and text cannot be compared
`replaceLinksWithEmbeds` attempts to compare the href and text of a link in a
percent-encoding agnostic way. This can be done by either percent-decoding the
href or percent-encoding the text. Both of these approaches can fail, but
`encodeURI` has fewer failure modes than `decodeURI`. Switch to using
`encodeURI` for this reason, handle failure, and add tests that would cover both
implementation approaches.

Fixes #4405