Remove unneeded GPU-acceleration class
It turns out this performance trick is no longer strictly needed for
the originally stated reason. The top bar remains cleanly fixed in
position when scrolling the page.

In Chrome the scrollable content area (the
.js-thread-list-scroll-root element) is now in its own compositing layer
anyway due to it having overflow: scroll set on it. This in turn
appears to force the browser to put the TopBar in its own compositor
layer in order to make it render above the scrollable area. When this
performance trick was originally added the whole sidebar body scrolled.