Split `useStore` hook into general and app-specific hooks
Split the `useStoreProxy` hook into a generic/base `useStore` hook that handles
wrapping a store created with `createStore`, and a `useSidebarStore` hook that
handles looking up the sidebar's main store via `useService('store')` and
passing it to the base hook.

This follows the existing separation of responsibilities between `createStore`
vs `createSidebarStore`, and would potentially allow us to split the generic
store infrastructure code into a separate package for use in other projects in
future.