Refine types of `config` parameters in annotator code
Various functions and constructors in the annotator code accepted a `config`
parameter with an uninformative `Record<string, any>` type.  Replace these types
with more specific ones that specify required / known optional properties
directly and use `Record<string, unknown>` for properties which are forwarded to
other contexts.

Also improve the documentation for the `getConfig` function which serves
as the entry point for reading configuration in the annotator.