Only check for flag changes before sending feature flags to annotator
Feature flags are now sent to the host frame, which never changes, so
there is no point in re-sending flags when guest frames change. In
future we'll need to figure out how to communicate feature flags to
guest frames and depending on whether the sidebar or host frame does
this, the sidebar may once again need to react to guest frames changing.